WIP: Flatten architecture. Remove v1 (sorry).
