--- openbb_agents/agents/v2/__init__.py
@@ -1,1 +1,0 @@
-from .agent import openbb_agent
